Clarify getPose behavior with visibile-blurred
The "poses may be reported" algorithm previously indicated that
poses could only be reported if the visibility state was "visible",
but that appears to be a mistake originating from the fact that
"visible-blurred" was added later. The description of
"visible-blurred" mentioned throttled head poses, so obviously the
intention was to allow some poses in that state, just not input
poses.

This change alters the "poses may be reported" algorithm to only
suppress poses when the visiblity state is "hidden", and adds an
explicit step to the "populate the pose" algorithm that prevents
input poses from being returned if the visibility state is
"visible-blurred"